step4 refactor - 패키징 변경  - view 패키지 생성 및 view 관련 클래스 이동
