+++ src/main/java/step2/race/controller/RaceController.java
@@ -4,6 +4,8 @@ import step2.car.domain.Car;
 import step2.car.domain.engine.RandomEngine;
 import step2.race.domain.Race;
 import step2.race.service.RaceService;
+import step2.view.InputView;
+import step2.view.OutputView;
 
 import java.util.List;
