+++ src/test/java/step2/race/controller/InputViewTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.MethodSource;
+import step2.view.InputView;
 
 import java.io.ByteArrayInputStream;
 import java.util.Arrays;
